inetmanet@fff5dfd manetrouting: Full takeover of MANET protocols from the INETMANET repo

+++ src/networklayer/manetrouting/batman/BatmanData.cc
@@ -1308,9 +1308,9 @@ int8_t Batman::send_udp_packet(cPacket *packet_buff, int32_t packet_buff_len, co
         return 0;
     }
     if (batman_if)
-        sendToIp(packet_buff, BATMAN_PORT, destAdd, BATMAN_PORT, 1, par("broadCastDelay"), batman_if->dev->ipv4Data()->getIPAddress().getInt());
+        sendToIp(packet_buff, BATMAN_PORT, destAdd, BATMAN_PORT, 1, par("broadCastDelay").doubleValue(), batman_if->dev->ipv4Data()->getIPAddress().getInt());
     else
-        sendToIp(packet_buff, BATMAN_PORT, destAdd, BATMAN_PORT, 1, par("broadCastDelay"), (Uint128)0);
+        sendToIp(packet_buff, BATMAN_PORT, destAdd, BATMAN_PORT, 1, par("broadCastDelay").doubleValue(), (Uint128)0);
     return 0;
 }
